refactor: refactored folders of sign in
