+++ src/application/rest/controllers/users/sign-in.controller.ts
@@ -12,7 +12,7 @@ import { ISendInfoLoggerProvider } from '@contracts/providers/logger/send-info-l
 import { Either, failure, success } from '@shared/utils/either.util';
 
 import { UseCase } from '@use-cases/_shared/use-case';
-import { SignInUseCaseDTO } from '@use-cases/users/sign-in/sign-in.use-case';
+import { SignInUseCaseDTO } from '@use-cases/users/sign-in.use-case';
 
 export class SignInController extends Controller<SignInControllerDTO.Parameters, SignInControllerDTO.Result> {
   constructor(
