+++ src/main/factories/use-cases/users/sign-in/credentials-sign-in.use-case.factory.ts
@@ -3,11 +3,9 @@ import { makeLoggerProvider } from '@factories/providers/logger-provider.factory
 import { makePasswordProvider } from '@factories/providers/password-provider.factory';
 import { makeUsersRepository } from '@factories/repositories/users-repository.factory';
 
+import { CredentialsSignInUseCaseDTO, CredentialsSignInUseCase } from '@use-cases/users/credentials-sign-in.use-case';
+
 import { UseCase } from '@domain/use-cases/_shared/use-case';
-import {
-  CredentialsSignInUseCaseDTO,
-  CredentialsSignInUseCase
-} from '@domain/use-cases/users/sign-in/credentials-sign-in.use-case';
 
 export const makeCredentialsSignInUseCase = (): UseCase<
   CredentialsSignInUseCaseDTO.Parameters,
